requirements.txt
nltk
app
newspaper3k
Flask-SQLAlchemy
Flask
sqlalchemy
db
jinja2
numpy
pandas
tensorflow